LayoutTests/imported/w3c/web-platform-tests/dom/nodes/moveBefore/moveBefore-shadow-root-expected.txt


FAIL moveBefore() is allowed in ShadowRoots (i.e., connected DocumentFragments) assert_equals: Original lastChild is now firstChild expected Element node <p>Child2</p> but got Element node <p>Child1</p>
